Add workaround for the issue with react events inside custom HTML components.
This is ugly hack to workaround issue with event propagation when react root is inside custom web component.
See https://github.com/facebook/react/issues/9242

This issue caused strange behaviour of platform components inside semantic-map popovers and Mirador annotation popovers.

Basically even `semantic-link` was not working properly there and instead of changing the content of the page dynamically it reloaded the whole page.
Or for example it was not possible to use `mp-event-trigger` in such context.

Signed-off-by: Artem Kozlov <artem@rem.sh>