Copy ontodia library source code into ResearchSpace.
Because we are going to use fork of the ontodia anyway, it is more convenient to have it tightly integrated into the ResearchSpace codebase.

Includes some changes to the library.

Improve export to png and svg.
* include all styles from the current HTML document because there is no way to extract reliably only css rules that are applicable to the diagram content
* embed fonts as data-uri
* embed original images as data-uri

Use ontodia prefixed label and type properties when getting entity info data.
That is necessary to make sure that when rdfs:label is actually not defined we don't add it to the model, the same for type. It is especially important in authoring mode.

Make sure that all entities are always typed as rdfs:Resource, even if not explicitly defined.

Don't decode dropped IRIs.
We have some resources that have encoded url params, so when we dropped them into ontodia they were decoded and become invalid.
Have no idea why that was needed in the first place, but it is definitely doesn't make any sense in ResearchSpace.

Allow editing of nodes in any state.
Previously it was possible to edit only nodes in expanded state.

Signed-off-by: Artem Kozlov <artem@rem.sh>