Change TemplateItem to properly work without a single root.
Previously many components that accept template as a parameter expected that the template should have a single root node.
That was because in React before v16 it was not possible to return array from render method.
But since React 16 it is now perfectly valid to do this.

Affects `semantic-table`, `semantic-query`, `semantic-tree`, `semantic-map`, `semantic-timeline`.

As an example see changes to `SemanticQuery` and `SemanticTimeline` help pages.

Signed-off-by: Artem Kozlov <artem@rem.sh>