Improve handling of anonymous user.
Previously every browser session created new session for anonymous user. So number of active session was equal to number of people that accessed the platform. That is especially painful in case of infinite session.

This change make sure that we have only one internal session for anonymous user, so we don't even need to set cookies in case of anonymous access.

Also it adds additional user `anonymous-user`, that can have permissions different from the actual `anonymous` user, but will have the same user IRI. This user can be useful in situation where anonymous user is readonly and one needs somehow to populate clipboard or other resource for it.

Signed-off-by: Artem Kozlov <artem@rem.sh>